RasberryApp/pyproject.toml
[project]
name = "rasb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pillow>=11.1.0",
    "pyqt6>=6.8.0",
    "structlog>=25.1.0",
]

[dependency-groups]
dev = [
    "rich>=13.9.4",
    "ruff>=0.9.3",
]
